Fix 500 error in Bill section and implement systemic robustness fixes in DAO, Model, and JSP layers

+++ src/main/java/com/oceanview/service/ReservationService.java
@@ -102,7 +102,7 @@ public class ReservationService {
         }
 
         long nights = res.getNumberOfNights();
-        return Payment.calculateBill(reservationId, nights, pricePerNight);
+        return Payment.calculateBill(reservationId, (double) nights, pricePerNight);
     }
 
     /**
